Reworked the analytical ray tracer to better handle the naked singularities
